+++ src/routes/ViewHoldingRoute.test.js
@@ -1,16 +1,27 @@
 import '../../test/jest/__mock__';
 
 import { MemoryRouter } from 'react-router-dom';
-import { render, screen } from '@folio/jest-config-stripes/testing-library/react';
+import {
+  render,
+  screen,
+} from '@folio/jest-config-stripes/testing-library/react';
+import {
+  QueryClient,
+  QueryClientProvider,
+} from 'react-query';
 
 import ViewHoldingRoute from './ViewHoldingRoute';
 
 jest.mock('../ViewHoldingsRecord', () => jest.fn().mockReturnValue('ViewHoldingsRecord'));
 
+const queryClient = new QueryClient();
+
 const wrapper = ({ children }) => (
-  <MemoryRouter>
-    {children}
-  </MemoryRouter>
+  <QueryClientProvider client={queryClient}>
+    <MemoryRouter>
+      {children}
+    </MemoryRouter>
+  </QueryClientProvider>
 );
 
 const renderViewHoldingRoute = (props = {}) => render(